upgraded to go 1.26, make update deps, divided s3client s3client interface and minio client into seperate files, added a little line in README.md, a pinch of fresh comments
